requirements.txt
black
coverage 
firebase_admin
Flask
Flask_Bootstrap
google-cloud-firestore
google-cloud-iam
google-cloud-logging
google-cloud-secret-manager
google-cloud-storage
ipaddr
mock-firestore
protobuf
PyJWT
pytest
pytest-mock
requests
sendgrid
tenacity
toml
waitress
Werkzeug
